06-advanced-integrations-and-demos/yocto-spdx-merge/pyproject.toml
[project]
name = "yocto-spdx-merge"
version = "0.1.0"
description = "Merge Yocto SPDX tar archives into a single SPDX 2.3 JSON document"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "spdx-tools>=0.8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
zstd = ["zstandard>=0.21"]

[project.scripts]
yocto-spdx-merge = "yocto_spdx_merge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yocto_spdx_merge"]
